Added dead eye consumption and kill reward

+++ common/src/main/java/com/namefix/network/payload/LevelDataPayload.java
@@ -6,8 +6,8 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 
-public record LevelDataPayload(int deadeyeSkill, int deadeyeLevel, int deadeyeXp) implements CustomPacketPayload {
-	public static final StreamCodec<RegistryFriendlyByteBuf, LevelDataPayload> CODEC = StreamCodec.composite(ByteBufCodecs.INT, LevelDataPayload::deadeyeSkill, ByteBufCodecs.INT, LevelDataPayload::deadeyeLevel, ByteBufCodecs.INT, LevelDataPayload::deadeyeXp, LevelDataPayload::new);
+public record LevelDataPayload(int deadeyeSkill, int deadeyeLevel, float deadeyeXp) implements CustomPacketPayload {
+	public static final StreamCodec<RegistryFriendlyByteBuf, LevelDataPayload> CODEC = StreamCodec.composite(ByteBufCodecs.INT, LevelDataPayload::deadeyeSkill, ByteBufCodecs.INT, LevelDataPayload::deadeyeLevel, ByteBufCodecs.FLOAT, LevelDataPayload::deadeyeXp, LevelDataPayload::new);
 
 	@Override
 	public Type<? extends CustomPacketPayload> type() {
